% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{AzureCredential}
\alias{AzureCredential}
\title{An Azure credential that selects and retains the first authentication method that succeeds.
Methods for requesting Azure access tokens.}
\usage{
AzureCredential
}
\description{
An Azure credential that selects and retains the first authentication method that succeeds.
Methods for requesting Azure access tokens.
}
\section{Methods}{

\subsection{Method \code{get_token}}{
Request an access token for one or more scopes.
}
}

