% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{default_azure_credential}
\alias{default_azure_credential}
\title{Get an access token using the default Azure credential chain.}
\usage{
default_azure_credential(scopes)
}
\arguments{
\item{scopes}{One or more Azure OAuth 2.0 scopes.}
}
\value{
A scalar character access token.
}
\description{
Get an access token using the default Azure credential chain.
}
