% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{default_azure_credential}
\alias{default_azure_credential}
\title{Create the default Azure credential chain.}
\usage{
default_azure_credential()
}
\description{
Create the default Azure credential chain.
}
